p31 2つのファイルに分割した
しかし、node --experimental-modulesがきかない
